+++ plugin-code/src/main/java/com/trivago/cluecumber/rendering/pages/pojos/pagecollections/AllStepsPageCollection.java
@@ -54,16 +54,24 @@ public class AllStepsPageCollection extends ScenarioSummaryPageCollection {
         return stepResultCounts.size();
     }
 
-    public String getMinimumTimeFromStep(final Step step){
-        return stepTimes.get(step).getMinimumTime();
+    public String getMinimumTimeFromStep(final Step step) {
+        return stepTimes.get(step).getMinimumTimeString();
     }
 
-    public String getMaximumTimeFromStep(final Step step){
-        return stepTimes.get(step).getMaximumTime();
+    public int getMinimumTimeScenarioIndexFromStep(final Step step) {
+        return stepTimes.get(step).getMinimumTimeScenarioIndex();
     }
 
-    public String getAverageTimeFromStep(final Step step){
-        return stepTimes.get(step).getAverageTime();
+    public String getMaximumTimeFromStep(final Step step) {
+        return stepTimes.get(step).getMaximumTimeString();
+    }
+
+    public int getMaximumTimeScenarioIndexFromStep(final Step step) {
+        return stepTimes.get(step).getMaximumTimeScenarioIndex();
+    }
+
+    public String getAverageTimeFromStep(final Step step) {
+        return stepTimes.get(step).getAverageTimeString();
     }
 
     /**
@@ -75,14 +83,17 @@ public class AllStepsPageCollection extends ScenarioSummaryPageCollection {
         if (reports == null) return;
         for (Report report : reports) {
             for (Element element : report.getElements()) {
+                int scenarioIndex = element.getScenarioIndex();
                 for (Step step : element.getSteps()) {
                     ResultCount stepResultCount = stepResultCounts.getOrDefault(step, new ResultCount());
                     updateResultCount(stepResultCount, step.getStatus());
                     stepResultCounts.put(step, stepResultCount);
-                    Times stepTimes = this.stepTimes.getOrDefault(step, new Times());
-                    stepTimes.addTime(step.getResult().getDuration());
-                    this.stepTimes.put(step, stepTimes);
                     addScenarioIndexByStatus(element.getStatus(), element.getScenarioIndex());
+                    Times stepTimes = this.stepTimes.getOrDefault(step, new Times());
+                    if (!step.isSkipped()) {
+                        stepTimes.addTime(step.getResult().getDuration(), scenarioIndex);
+                    }
+                    this.stepTimes.put(step, stepTimes);
                 }
             }
         }