+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Back to [Readme](README.md).
 
+## [1.9.0] - 2019-04-10
+
+### Added
+
+* Minimum and maximum step times on `All Steps` page links to the enclosing scenario (#152)
+* New `failScenariosOnPendingOrUndefinedSteps` configuration property for marking scenarios as failed when they contain
+steps with status `pending` or `undefined` (default value is `false`) (#74)
+
+### Changed
+
+* Minimum and maximum step times are now only calculated for non-skipped steps
+
 ## [1.8.1] - 2019-04-03
 
 ### Fixed
@@ -394,6 +406,7 @@ Back to [Readme](README.md).
 
 Initial project version on GitHub and Maven Central.
 
+[1.9.0]: https://github.com/trivago/cluecumber-report-plugin/tree/1.9.0
 [1.8.1]: https://github.com/trivago/cluecumber-report-plugin/tree/1.8.1
 [1.8.0]: https://github.com/trivago/cluecumber-report-plugin/tree/1.8.0
 [1.7.3]: https://github.com/trivago/cluecumber-report-plugin/tree/1.7.3